add lyricsHandle and rename to songsHandle.js Separate parseLyrics into lyricsHandle.js
