+++ src/modules/commit-info.js
@@ -1,1 +1,1 @@
-export const COMMIT_ID = '55803cdf3445'
+export const COMMIT_ID = '1ae236605ffb'
